+++ src/main/kotlin/next/step/lotto/domain/LottoWinningCount.kt
@@ -1,4 +1,4 @@
-package next.step.lotto
+package next.step.lotto.domain
 
 enum class LottoWinningCount(val matchCount: Int, val winnings: Int) {
     THREE(3, 5000),
@@ -6,7 +6,7 @@ enum class LottoWinningCount(val matchCount: Int, val winnings: Int) {
     FIVE(5, 1500000),
     SIX(6, 2000000000),
     NONE(0, 0);
-    
+
     companion object {
         fun from(matchCount: Int): LottoWinningCount = values().find { it.matchCount == matchCount } ?: NONE
     }
