+++ src/main/kotlin/next/step/lotto/domain/Lottos.kt
@@ -1,10 +1,10 @@
-package next.step.lotto
+package next.step.lotto.domain
 
 @JvmInline
 value class Lottos(val lottos: Set<Lotto>) : Set<Lotto> by lottos {
     fun match(winningNumbers: LottoWinningNumbers): LottoWinningStat =
         LottoWinningStat.of(lottos.groupingBy { it.match(winningNumbers) }.eachCount())
-    
+
     companion object {
         fun buy(payment: Int): Lottos {
             val lottos = mutableSetOf<Lotto>()
