+++ src/test/kotlin/next/step/lotto/domain/LottosTest.kt
@@ -1,4 +1,4 @@
-package next.step.lotto
+package next.step.lotto.domain
 
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.datatest.withData
